Fix clippy warnings across codebase
- Remove unused fuser::MountOption import and fix unused must-use
  in src/commands/mount.rs
- Simplify state initialization in mount_session.rs
- Allow unreachable_patterns for libc aliases in workspace.rs
- Allow dead_code for OpenFileState::Closed variant
- Use div_ceil in mount/attr.rs
- Convert match to matches! macro in runtime/execution.rs
- Fix manual ok() implementations in cache/mod.rs
- Remove unnecessary casts in commands/cp.rs
- Simplify map_or in fs/ops.rs
- Remove unneeded return in workspace.rs
- Remove explicit .into_iter() in shell/repl.rs
- Add type alias for complex tuple in storage/sqlite.rs
- Remove needless borrows in storage/sled.rs
- Use sort_by_key instead of sort_by in storage/sled.rs and lmdb.rs
- Replace needless range loops with iterators in storage/sled.rs and lmdb.rs

+++ src/commands/cp.rs
@@ -47,7 +47,7 @@ pub fn run(args: CpArgs, output: &Output, vault: Option<String>) -> Result<()> {
     let (logical_size, file_count) = if entry.is_dir() {
         calculate_tree_size(&fs, &args.source)?
     } else {
-        (entry.size as u64, 1usize)
+        (entry.size, 1usize)
     };
 
     // Copies reuse existing content blobs, so they only consume file-count quota.
@@ -88,7 +88,7 @@ fn calculate_tree_size(fs: &FileSystem, path: &str) -> Result<(u64, usize)> {
     let entry = fs.get_entry(path)?;
 
     if !entry.is_dir() {
-        return Ok((entry.size as u64, 1));
+        return Ok((entry.size, 1));
     }
 
     let mut total_size = 0u64;
@@ -107,7 +107,7 @@ fn calculate_tree_size(fs: &FileSystem, path: &str) -> Result<(u64, usize)> {
             total_size += sub_size;
             file_count += sub_count;
         } else {
-            total_size += child.size as u64;
+            total_size += child.size;
             file_count += 1;
         }
     }
